conversaciones.py
from flask import Flask, request
from conversaciones import iniciar_conversacion, cerrar_conversacion, obtener_contexto_y_historial, guardar_contexto_y_historial
from mi_modulo_del_bot import tu_funcion_del_bot

app = Flask(__name__)

@app.route('/whatsapp', methods=['POST'])
def process_message():
    # Obtener el mensaje y el número de teléfono del remitente
    message = request.form.get('Body')
    phone_number_string = request.form.get('From')

    # Procesar número de teléfono y obtener contexto e historial
    formatted_phone_number = process_phone_number(phone_number_string)
    contexto, historial_conversacion = obtener_contexto_y_historial(formatted_phone_number)

    # Iniciar la conversación si es la primera interacción
    if contexto is None:
        iniciar_conversacion(formatted_phone_number)
        contexto = {}  # Definir el contexto inicial

    # Llamar a la función del bot y obtener la respuesta
    respuesta = tu_funcion_del_bot(message, contexto)

    # Actualizar el contexto y el historial
    historial_conversacion.append({"mensaje": message, "remitente": "Usuario"})
    historial_conversacion.append({"mensaje": respuesta, "remitente": "Bot"})
    guardar_contexto_y_historial(formatted_phone_number, contexto, historial_conversacion)

    # Enviar la respuesta al usuario de WhatsApp
    enviar_respuesta_a_whatsapp(respuesta, formatted_phone_number)

    return '', 200

def enviar_respuesta_a_whatsapp(respuesta, formatted_phone_number):
    # Código para enviar la respuesta al usuario de WhatsApp
    pass

def process_phone_number(phone_number_string):
    # Código para procesar el número de teléfono a formato E.164
    pass

if __name__ == '__main__':
    app.run(port=65535)
