requirements.txt
config
Flask
twilio
openai
spacy
python-dotenv
nltk
emoji
phonenumbers
zapier
